@@ -39,79 +39,78 @@
   "peerDependencies": {
     "@skbkontur/react-icons": ">=3 <=4",
     "@skbkontur/react-ui": ">=2 <3",
-    "react": ">=15 <=16",
-    "react-dom": ">=15 <=16",
+    "react": ">=16 <=17",
+    "react-dom": ">=16 <=17",
     "react-router": ">=4 <=5",
     "react-router-dom": ">=4 <=5"
   },
   "dependencies": {
     "@skbkontur/react-stack-layout": "^1.0.3",
-    "@skbkontur/react-ui-validations": "^1.1.4",
+    "@skbkontur/react-ui-validations": "^1.4.0",
     "copy-to-clipboard": "^3.3.1",
-    "decimal.js": "^10.2.0",
-    "lodash": "^4.17.15",
-    "moment": "^2.24.0",
-    "qs": "^6.9.3",
-    "whatwg-fetch": "^3.0.0"
+    "decimal.js": "^10.2.1",
+    "lodash": "^4.17.20",
+    "moment": "^2.29.1",
+    "qs": "^6.9.6",
+    "whatwg-fetch": "^3.5.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.9.5",
-    "@babel/plugin-transform-runtime": "^7.9.0",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-react": "^7.9.4",
-    "@babel/preset-typescript": "^7.9.0",
-    "@babel/runtime": "^7.9.2",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.2",
-    "@skbkontur/react-icons": "^4.3.0",
-    "@skbkontur/react-ui": "2.0.0",
-    "@storybook/addon-actions": "^6.0.0-beta.32",
-    "@storybook/addons": "^6.0.0-beta.32",
-    "@storybook/react": "^6.0.0-beta.32",
-    "@types/chai": "^4.2.11",
-    "@types/lodash": "^4.14.150",
-    "@types/mocha": "^7.0.2",
-    "@types/qs": "^6.9.1",
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.7",
-    "@types/react-router": "^5.1.7",
-    "@types/react-router-dom": "^5.1.5",
+    "@babel/core": "^7.12.10",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
+    "@babel/plugin-transform-runtime": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-react": "^7.12.10",
+    "@babel/preset-typescript": "^7.12.7",
+    "@babel/runtime": "^7.12.5",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
+    "@skbkontur/react-icons": "^4.4.0",
+    "@skbkontur/react-ui": "^2.13.0",
+    "@storybook/addon-actions": "^6.1.15",
+    "@storybook/addons": "^6.1.15",
+    "@storybook/react": "^6.1.15",
+    "@types/chai": "^4.2.14",
+    "@types/lodash": "^4.14.168",
+    "@types/mocha": "^8.2.0",
+    "@types/qs": "^6.9.5",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-router": "^5.1.11",
+    "@types/react-router-dom": "^5.1.7",
     "@types/storybook-react-router": "^1.0.1",
-    "@typescript-eslint/eslint-plugin": "^3.3.0",
-    "@typescript-eslint/parser": "^3.3.0",
-    "babel-loader": "^8.1.0",
+    "@typescript-eslint/eslint-plugin": "^4.14.1",
+    "@typescript-eslint/parser": "^4.14.1",
+    "babel-loader": "^8.2.2",
     "chai": "^4.2.0",
-    "copyfiles": "^2.2.0",
-    "css-loader": "^3.5.3",
-    "eslint": "^7.3.0",
-    "eslint-config-prettier": "^6.11.0",
+    "copyfiles": "^2.4.1",
+    "css-loader": "^5.0.1",
+    "eslint": "^7.18.0",
+    "eslint-config-prettier": "^7.2.0",
     "eslint-import-resolver-node": "^0.3.4",
-    "eslint-plugin-import": "^2.21.2",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.0",
-    "less": "^3.11.1",
-    "less-loader": "^6.0.0",
-    "mocha": "^7.1.2",
-    "nerdbank-gitversioning": "^3.1.74",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-react": "^7.22.0",
+    "less": "^4.1.0",
+    "less-loader": "^7.3.0",
+    "mocha": "^8.2.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.0.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-refresh": "^0.8.3",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
+    "prettier": "^2.2.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-refresh": "^0.9.0",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "rimraf": "^3.0.2",
     "storybook-react-router": "^1.0.8",
-    "style-loader": "^1.2.0",
-    "stylelint": "^13.6.1",
+    "style-loader": "^2.0.0",
+    "stylelint": "^13.9.0",
     "stylelint-config-css-modules": "^2.2.0",
     "stylelint-config-standard": "^20.0.0",
-    "tslib": "^2.0.0",
-    "typescript": "^3.9.5",
-    "url-loader": "^4.1.0",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
+    "tslib": "^2.1.0",
+    "typescript": "^4.1.3",
+    "url-loader": "^4.1.1",
+    "webpack": "^4.46.0",
+    "webpack-cli": "^4.4.0",
+    "webpack-dev-server": "^3.11.2"
   }
 }
